requirements.txt
mysql-connector
requests==2.25.1
Flask_Cors==3.0.10
pandas==1.2.4
pytz==2021.1
Flask==2.0.0
SQLAlchemy==1.4.15
numpy==1.20.2
yfinance==0.1.63
python_dateutil==2.8.2
